DepthFunc 추가. 뎁스 테스트 켜고 끄기 추가. ClearColor 및 ClearDepth 추가.

+++ SoftwareRenderer/WinMain.cpp
@@ -153,7 +153,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 			RECT rect;
 			GetClientRect( hWnd, &rect );
 			renderer.reset( new Renderer::GdiRenderer( hWnd, rect.right, rect.bottom ) );
-			renderer->Viewport( 0, 0, (float)rect.right, (float)rect.bottom );
+			renderer->Viewport( 0, 0, rect.right, rect.bottom );
 			renderer->Perspective( 45.f, (float)rect.right / (float)rect.bottom, 0.1f, 100.f );
             scene = SceneMenu::Select( renderer, hWnd, ID_SCENE_DEFAULT );
 			break;
